backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==6.0.0
Flask-Limiter==3.6.0
flask-swagger-ui==4.11.1

# Database
psycopg2-binary==2.9.9
redis==5.0.4
SQLAlchemy==2.0.23

# Security
cryptography==42.0.7
PyJWT==2.10.1

# QR & Image
qrcode[pil]==8.2
Pillow==10.4.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.1.0
click==8.1.8
python-dateutil==2.9.0.post0

# Production
gunicorn==23.0.0
gevent==25.5.1

# Testing (dev only)
pytest==8.2.1
pytest-flask==1.3.0
pytest-cov==6.1.1